backend:refactor split friendRequest views
